add more functions
add more functions

+++ NBC.Web/Areas/Admin/Controllers/UsersController.cs
@@ -12,6 +12,9 @@ using System.Threading.Tasks;
 using NBC.Models;
 using System.Net;
 using System.Collections;
+using System.Security.Cryptography;
+using System.Text;
+using System.Data;
 
 namespace NBC.Web.Areas.Admin.Controllers
 {
@@ -26,20 +29,39 @@ namespace NBC.Web.Areas.Admin.Controllers
         private RoleService RoleService;
         private UserInRoleService UserInRoleService;
         private SVService SVService;
+        private UnitService UnitService;
+       
         public UsersController(UserService UserService, SettingService settingService, RoleService roleService, UserInRoleService userInRoleService
-            , SVService svService)
+            , SVService svService, UnitService unitService)
         {
             this.UserService = UserService;
             this.settingService = settingService;
             this.RoleService = roleService;
             this.UserInRoleService = userInRoleService;
             this.SVService = svService;
+            this.UnitService = unitService;
         }
         public ActionResult Index()
         {
             return View(UserService.All());
         }
-        public ActionResult Details(int? id)
+        public async Task<ActionResult> Details(int? id)
+        {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            User user = UserService.Find(id);
+            var thisUser = await UserManager.FindByNameAsync(user.UserName);
+            string password = thisUser.PasswordHash;
+            //  ViewBag.Password = base64Decode(password);
+            if (user == null)
+            {
+                return HttpNotFound();
+            }
+            return PartialView(user);
+        }
+        public ActionResult Edit(int? id)
         {
             if (id == null)
             {
@@ -52,11 +74,24 @@ namespace NBC.Web.Areas.Admin.Controllers
             }
             return PartialView(user);
         }
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult Edit(User user)
+        {
+            if (ModelState.IsValid)
+            {
+                //db.Entry(unit).State = EntityState.Modified;               
+                UserService.SetModified(user);
+                UserService.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            return View(user);
+        }
         public ActionResult Create()
         {
 
 
-            return View();
+            return PartialView();
         }
         [HttpPost]
         [AllowAnonymous]
@@ -82,19 +117,18 @@ namespace NBC.Web.Areas.Admin.Controllers
                         // var callbackUrl = Url.Action("ConfirmEmail", "Account", new { userId = user.Id, code = code }, protocol: Request.Url.Scheme);
                         // await UserManager.SendEmailAsync(user.Id, "Confirm your account", "Please confirm your account by clicking <a href=\"" + callbackUrl + "\">here</a>");
                         // UserService.Add(model);
+                        model.isEnable = true;
+
                         User thisUser = model;
                         Role role = RoleService.GetRoleByName("Subscribers");
                         UserInRole inRole = new UserInRole();
-                        inRole.User = thisUser;
-                        inRole.Role = role;
-                        NBC.Models.SV thisSV = SVService.Find(9);
-                        //inRole.WorkAt = thisSV;
+                        inRole.User = thisUser;                        
                         inRole.isEnable = true;
-                        //inRole.CreatedDate = DateTime.Now;
-                        //inRole.ModifiedDate = DateTime.Now;
-
-
-                        UserInRoleService.Add(inRole);
+                        
+                        inRole.Role = role;
+                        inRole = UserInRoleService.Add(inRole);
+                       
+                        UserInRoleService.SaveChanges();
                         return RedirectToAction("Index");
 
                     }
@@ -137,67 +171,84 @@ namespace NBC.Web.Areas.Admin.Controllers
         // POST: Admin/Years/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public async Task<ActionResult> DeleteConfirmed(int id)
         {
             User user = UserService.Find(id);
 
             try
             {
-               
-                var thisUser = new ApplicationUser { UserName = user.UserName, Email = user.Email };
-                UserManager.Delete(thisUser);
+
+                var thisUser = await UserManager.FindByNameAsync(user.UserName);
+                if (user == null)
+                {
+                    return HttpNotFound();
+                }
+                var result = await UserManager.DeleteAsync(thisUser);
+                //   var roles =  UserInRoleService.GetUserInRoleByUserId(user.Id);
+                //foreach(var role in roles)
+                //{
+                //    UserInRoleService.Remove(role);
+                //    UserInRoleService.SaveChanges();
+                //}
+
                 UserService.Remove(user);
                 UserService.SaveChanges();
 
                 return RedirectToAction("Index");
             }
-            catch (Exception)
+            catch (Exception ex)
             {
 
                 return PartialView(user);
             }
 
         }
-        public ActionResult AddUserInRole(int? id)
+        public ActionResult AddUserInRole(int id)
         {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
 
-            List<UserInRole> userInRole = UserInRoleService.GetUserInRoleByUserId(id);
-            if (userInRole == null)
+            User user = UserService.GetUserById(id);
+
+            // List<UserInRole> userInRole = UserInRoleService.GetUserInRoleByUserId(id);
+            if (user == null)
             {
                 return HttpNotFound();
             }
             var role = RoleService.All().ToList();
             ViewBag.role = role;
-            return PartialView(userInRole);
+            return PartialView(user);
         }
         [HttpPost]
-        public ActionResult SaveUserInRole(int id, string UserRole_Id)
+        public ActionResult SaveUserInRole(int id, string UserRole_Id,string WID,string roleName)
         {
 
-            List<UserInRole> userInRole = UserInRoleService.GetUserInRoleByUserId(id);
-            if (userInRole == null)
+            UserInRole userInRole = UserInRoleService.GetUserInRoleByUserIdAndUserRoleId(id, Convert.ToInt32(UserRole_Id));
+            if (userInRole != null)
             {
                 return HttpNotFound();
             }
-
-            UserInRole inRole = new UserInRole();
-            inRole.User.Id = id;
-            inRole.Role.Id = Convert.ToInt32(UserRole_Id);
-            inRole = UserInRoleService.Add(inRole);
-            if (inRole == null)
+           // int WID = ViewBag.WID;
+            userInRole = new UserInRole();
+            Role role = RoleService.GetUserRoleById(Convert.ToInt32(UserRole_Id));
+            User user = UserService.GetUserById(id);
+           
+           
+            userInRole.User = user;
+            userInRole.Role = role;
+            if (roleName == "SV")
             {
-                return Content("UserName ที่ท่านกำลังกำหนดสิทธิ์ มีสิทธิ์นี้ในระบบแล้ว");
+                NBC.Models.SV sv = SVService.GetWorkAtSV(Convert.ToInt32(WID));
+                userInRole.WorkAt = sv;
             }
-            else
-            {
-                return Content("OK");
+            else if (roleName == "Unit") {
+                NBC.Models.Unit unit = UnitService.GetWorkAtUnit(Convert.ToInt32(WID));
+                userInRole.WorkAt = unit;
             }
-
-
+            
+            userInRole = UserInRoleService.Add(userInRole);
+            UserInRoleService.SaveChanges();
+            
+            return Content("OK");
+           
         }
         [HttpPost]
         public ActionResult DeleteUserInRole(int id)
@@ -210,9 +261,52 @@ namespace NBC.Web.Areas.Admin.Controllers
             }
             else
             {
-                userInRole = UserInRoleService.Remove(userInRole);
-                return Content("Ok");
+                try
+                {
+                    UserInRoleService.Remove(userInRole);
+                    UserInRoleService.SaveChanges();
+
+                    return Content("OK");
+                }
+                catch (Exception ex)
+                {
+
+                    return Content(ex.ToString());
+                }
+
             }
+
+        }
+        [HttpPost]
+        public ActionResult SetEnableUser(int? id)
+        {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            try
+            {
+                User user = UserService.Find(id);
+                if (user.isEnable == true)
+                {
+                    user.isEnable = false;
+                }
+                else
+                {
+                    user.isEnable = true;
+                }
+                UserService.SetModified(user);
+                UserService.SaveChanges();
+                return RedirectToAction("Index");
+                // return this.Json(new { success = true });
+            }
+            catch (System.Exception)
+            {
+
+                return this.Json(new { success = false });
+            }
+
+
 
         }
         public ApplicationUserManager UserManager
@@ -244,6 +338,29 @@ namespace NBC.Web.Areas.Admin.Controllers
                 ModelState.AddModelError("", error);
             }
         }
+        [HttpGet]
+        public ActionResult DDLWorkAtList(int id)
+        {
+            Role role = RoleService.Find(id);
+            if (role == null)
+            {
+                return HttpNotFound();
+            }
+            switch (role.RoleName)
+            {
+                case "SV":
+                    List<NBC.Models.SV> sv = SVService.All().ToList();
+                    return Json(sv, JsonRequestBehavior.AllowGet);
+
+                case "Unit":
+                    List<NBC.Models.Unit> unit = UnitService.All().ToList();
+                    return Json(unit, JsonRequestBehavior.AllowGet);
+                default:
+                    return View();
+                    ;
+            }
+        }
+
 
     }
 }
